// entraverse/types/product.types.ts
export interface ProductVariantGroup {
  name: string;
  options: string[];
}

export interface ProductVariantPricingRow {
  sku?: string;
  sku_seller?: string;
  variant_code?: string;
  label?: string;
  options?: Record<string, string>;
  stock?: number;
  item_weight?: number;
  offline_price?: number;
  entraverse_price?: number;
  tokopedia_price?: number;
  shopee_price?: number;
}